requirements.txt
torchgeometry==0.1.2
tqdm==4.51.0
pycocotools==2.0.2
torch==1.6.0
torchvision==0.7.0
chumpy==0.69
matplotlib==3.3.2
torch_geometric==1.6.1
scipy==1.5.3
numpy==1.19.2
trimesh==3.8.12
opencv_python==4.4.0.46
ext==0.1.0
geometry==0.0.23
Pillow==8.3.2
pywavefront==1.3.3
scipy==1.5.3
vispy==0.5.3
Pillow==8.3.2
